worked on the bottom navigation bar
